// client/src/features/authentication/types/user.type.ts
export type UserType = {
  _id: string
  email: string
  password: string
  role: string
  banner: string
  avatar: string
  name: string
  nickName: string
  phoneNumber: string
  address: string
  birthday: string
  gender: string
  work: string
  education: string
  skills: string
  bio: string

  createdAt: string
  updatedAt: string

  isFollowing: boolean
  total_follower: number
}